refactor: move setup and vault entrypoints to TypeScript
